+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 ## Release Notes
 
-### TBD
+### v0.9.7 (2015-10-07)
 
 * Add `ignoreMissingHandler` option to swagger-router middleware to allow the API author to dictate how to handle
 missing handlers *(Issue #274)*